V1.2 - Playing using scrren
- Identify Coin
- Identify Boost
- Identify Orange Box
- Jump
- Boost
